Drop the conversion on a value that was already signed
CI's lint job caught what mine could not: unconvert on int64(fs.Bsize),
which Statfs_t already declares as int64 on both platforms this builds for.
Bavail is unsigned and still needs converting, which is now said in a
comment rather than implied by the symmetry of writing both.

The reason it reached main is worth recording: golangci-lint on this
machine refuses to run against a go.mod targeting a newer Go than it was
built with, so every check before this one was go vet and gofmt standing in
for a linter that never ran.  Rebuilt it at the version CI pins, under the
toolchain this module targets, and it reproduces CI exactly: one issue
before, none after.

+++ internal/audit/audit.go
@@ -224,7 +224,8 @@ func (l *Log) Unwritable() string {
 		log.Printf("cannot measure free space under %s: %v", l.config.LogPath, err)
 		return ""
 	}
-	free := int64(fs.Bavail) * int64(fs.Bsize)
+	// Bavail is unsigned and Bsize is not, which is why only one is converted.
+	free := int64(fs.Bavail) * fs.Bsize
 	if want := int64(l.config.MaxRecordBytes); free < want {
 		return fmt.Sprintf("%s has %d bytes free and one record may need %d",
 			filepath.Dir(l.config.LogPath), free, want)
